legacyWatcher: layer configuration dialog is working and hooked in. Updates the layer configuration on the fly. Still a little buggy and not all settings are working, but the structure is all there so it's just a matter of filling it in.  There are also a few general bugs to be fixed, but I'm out of time and want to check this in before I go on vacation.

+++ src/libwatcher/edgeDisplayInfo.h
@@ -52,7 +52,7 @@ namespace watcher
 
             std::string label;                  ///< written next to the edge, not used often
             std::string labelFont; 
-            double labelPointSize; 
+            float labelPointSize; 
             watcher::Color labelColor;
 
             bool loadConfiguration(const watcher::event::GUILayer &layer);
